pyscraper/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
elasticsearch==7.11.0
idna==2.10
numpy==1.20.1
pandas==1.2.2
pydantic==1.8.1
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
typing-extensions==3.7.4.3
urllib3==1.26.3
